Extract integration tests for unique generation
